[CommCoreModule] Change `Message.h:time` from `int` to `long long`
Summary:
`std::stoi` was failing because the timestamps were too big.

`time` column has INTEGER affinity but we don't need to change anything on sqlite side:
"INTEGER. The value is a signed integer, stored in 1, 2, 3, 4, 6, or 8 bytes depending on the magnitude of the value." (via docs)

Test Plan: `std::stoll` succeeds, timestamp fits in long long, timestamp persisted in sqlite successfully: https://blob.sh/atul/e34f.png

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2007

+++ native/cpp/CommonCpp/DatabaseManagers/entities/Message.h
@@ -11,7 +11,7 @@ struct Message {
   int type;
   int future_type;
   std::string content;
-  int time;
+  int64_t time;
 };
 
 } // namespace comm
